fix(DTFS2-6621): update tests
